Split AccountRow component into presentation and logical container
